Extract pw common functions
